+++ app/config/command.config.php
@@ -7,18 +7,28 @@ return array(
     "class"     => "HelpCommand",
     "fn"        => "println",
   ],
+
   [
     "cmd"       => ["-v", "--version"],
     'mode'      => "full",
     "class"     => "HelpCommand",
     "fn"        => "versionCek",
   ],
+
+  [
+    "cmd"       => "--list",
+    'mode'      => "full",
+    "class"     => "HelpCommand",
+    "fn"        => "commandList",
+  ],
+
   [
     "cmd"       => "make",
     "mode"      => "start",
     "class"     => "MakerCommand",
     "fn"        => "switcher",
   ],
+
   [
     "cmd"       => "serve",
     "mode"      => "full",