+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- CLI support multy echo using prints(), and new line & tabs)
+- MyPDO support new constructor with static function
+
+### Fixed
+- CLI make:model/models not replace table name and/or table entry (columns)
+- CLI make:models class name not same with file name
+- Services handle class funtion if not return as array (default is empty array)
 
 ## [0.1.2] - 2021-03-21
 ### Added